test: refactor mq_utils tests
- move tests from test_mq_utils_read.c to test_mq_utils.c
- delete test_mq_utils_read.c
